+++ App.tsx
@@ -1,11 +1,11 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import HomePage from './src/Home';
+import Home from './src/Home';
 
 export default function App() {
   return (
     <View style={styles.container}>
-      <HomePage />
+      <Home />
       <StatusBar style="auto" />
     </View>
   );
